w:Dockerfile.debian: ->bullseye + mmbdebstrap
In order to build debian packages on debian machines, we need a newer
suite of mmdebstrap for srebuild. Update to the up-and-coming debian
stable (bullseye, now testing) to use a newer (now also explicitly
installed) mmdebstrap.

Lastly, to keep things consistent between Dockerfiles, use ENTRYPOINT
rather than CMD (this also keeps things a little neater when issuing
workers for testing)

Note: we are using bullseye as a tag for the container stage (and not
testing) to ensure a nicer handoff in the coming weeks.