ENH: add minimal instructions to run debian builds
For the sake of my memory, and because it could be helpful to those
trying to manage workers as containers, document the process a little
bit in its own readme. It goes without saying, the debian docker
container needs CAP_SYS_ADMIN, and that's needs to be adequately
called out.