Implement support for storing compressed build logs.
Logs are now stored in the database as zstd-compressed blobs by default. Clients requesting build logs may
be served the raw zstd-compressed data if they indicate support for it via the Accept-Encoding header; otherwise,
the logs are decompressed and served as plaintext.

Additional compression may be applied by the Actix compression middleware on plaintext logs if the client supports
compressions other than zstd.