Fix prioritization of newly-enqueued packages.
Firstly, the whole queue was being overwritten. Now, only packages that currently do not have a queue item will be enqueued.
Second, the priority was the same for good and bad packages, meaning it was less and less likely that an unknown package would get picked up. Now, packages without a queue item that are known to build BAD will have a lower priority than unknown packages.